#22 moved streaming related tests to own test-suite
