+++ src/test/scala/mockws/MockWSTest.scala
@@ -102,33 +102,6 @@ class MockWSTest extends FunSuite with Matchers with PropertyChecks {
   }
 
 
-  test("mock WS simulates a streaming") {
-
-    def testedController(ws: WSClient) = Action.async {
-      ws.url("/").stream().map { resp =>
-        Result(
-          header = ResponseHeader(resp.headers.status, resp.headers.headers.mapValues(_.head)),
-          body = HttpEntity.Streamed(resp.body, None, None))
-      }
-    }
-
-    val ws = MockWS {
-      case (GET, "/") => Action {
-        val body: Source[ByteString, _] = Source(Seq("first", "second", "third").map(ByteString.apply))
-        Result(
-          header = ResponseHeader(201, Map("x-header" -> "x-value")),
-          body = HttpEntity.Streamed(body, None, None))
-      }
-    }
-    import ws.materializer
-
-    val response = testedController(ws).apply(FakeRequest())
-    status(response) shouldEqual CREATED
-    contentAsString(response) shouldEqual "firstsecondthird"
-    header("x-header", response) shouldEqual Some("x-value")
-    ws.close()
-  }
-
   test("mock WS can produce JSON") {
     val ws = MockWS {
       case (GET, "/json") => Action {
@@ -256,34 +229,6 @@ class MockWSTest extends FunSuite with Matchers with PropertyChecks {
     ws.close()
   }
   
-  test("mock WS supports method in stream") {
-    def testedController(ws: WSClient) = Action.async {
-      ws.url("/").withMethod("POST").stream().map { resp =>
-        Result(
-          header = ResponseHeader(resp.headers.status, resp.headers.headers.mapValues(_.head)),
-          body = HttpEntity.Streamed(resp.body, None, None)
-        )
-      }
-    }
-
-    val ws = MockWS {
-      case (POST, "/") => Action {
-        val body: Source[ByteString, _] = Source(Seq("first", "second", "third").map(ByteString.apply))
-        Result(
-          header = ResponseHeader(201, Map("x-header" -> "x-value")),
-          body = HttpEntity.Streamed(body, None, None)
-        )
-      }
-    }
-    import ws.materializer
-
-    val response = testedController(ws).apply(FakeRequest())
-    status(response) shouldEqual CREATED
-    contentAsString(response) shouldEqual "firstsecondthird"
-    header("x-header", response) shouldEqual Some("x-value")
-    ws.close()
-  }
-
   test("should not raise NullPointerExceptions on method chaining") {
     val ws = MockWS {
       case (GET, "/get") => Action { Ok("get ok") }
@@ -300,21 +245,4 @@ class MockWSTest extends FunSuite with Matchers with PropertyChecks {
     ws.close()
   }
 
-  test("should pass through all elements of a Source") {
-    val content = Source(Seq("hello, ", "world").map(ByteString(_)))
-
-    val ws = MockWS {
-      case (GET, "/get") ⇒ Action {
-        Result(
-          header = ResponseHeader(200),
-          body = HttpEntity.Streamed(content, None, None)
-        )
-      }
-    }
-
-    await(ws
-      .url("/get")
-      .get()).body shouldEqual "hello, world"
-    ws.close()
-  }
 }
